# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "HealthController_check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The Health Check is successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {},
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The Health Check is not successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "info": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "error": {
                      "type": "object",
                      "example": {
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      },
                      "nullable": true
                    },
                    "details": {
                      "type": "object",
                      "example": {
                        "database": {
                          "status": "up"
                        },
                        "redis": {
                          "status": "down",
                          "message": "Could not connect"
                        }
                      },
                      "additionalProperties": {
                        "type": "object",
                        "required": [
                          "status"
                        ],
                        "properties": {
                          "status": {
                            "type": "string"
                          }
                        },
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/produtores": {
      "post": {
        "operationId": "ProdutoresController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarProdutorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutorDto"
                }
              }
            }
          },
          "400": {
            "description": "O Documento ou o nome informado não é válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Já existe um Produtor com esse Documento.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Registra um Produtor.",
        "tags": [
          "Produtores"
        ]
      }
    },
    "/produtores/{id}": {
      "get": {
        "operationId": "ProdutoresController_buscar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutorDetalhadoDto"
                }
              }
            }
          },
          "404": {
            "description": "Não existe Produtor com esse identificador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Recupera um Produtor e suas Propriedades, com o Documento mascarado.",
        "tags": [
          "Produtores"
        ]
      }
    },
    "/propriedades": {
      "post": {
        "operationId": "PropriedadesController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarPropriedadeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropriedadeDto"
                }
              }
            }
          },
          "400": {
            "description": "A entrada não é válida, ou a soma da Área Agricultável com a Área de Vegetação passa da Área Total.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Não existe Produtor com esse identificador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Registra uma Propriedade em nome de um Produtor.",
        "tags": [
          "Propriedades"
        ]
      },
      "get": {
        "operationId": "PropriedadesController_listar",
        "parameters": [
          {
            "name": "pagina",
            "required": false,
            "in": "query",
            "description": "A página pedida. A primeira é a de número um.",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "tamanho",
            "required": false,
            "in": "query",
            "description": "Quantos registros por página, no máximo 100.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropriedadesPaginaDto"
                }
              }
            }
          }
        },
        "summary": "Lista as Propriedades, da mais recente para a mais antiga.",
        "tags": [
          "Propriedades"
        ]
      }
    },
    "/propriedades/{id}": {
      "put": {
        "operationId": "PropriedadesController_editar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditarPropriedadeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropriedadeDto"
                }
              }
            }
          },
          "400": {
            "description": "A entrada não é válida, ou a soma da Área Agricultável com a Área de Vegetação passa da Área Total.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Não existe Propriedade com esse identificador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Atualiza a localização e as áreas de uma Propriedade.",
        "tags": [
          "Propriedades"
        ]
      },
      "delete": {
        "operationId": "PropriedadesController_excluir",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "A Propriedade foi excluída."
          },
          "404": {
            "description": "Não existe Propriedade com esse identificador.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Exclui uma Propriedade, e com ela seus Plantios.",
        "tags": [
          "Propriedades"
        ]
      }
    },
    "/safras": {
      "post": {
        "operationId": "SafrasController_criar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarSafraDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafraDto"
                }
              }
            }
          },
          "400": {
            "description": "O ano informado não serve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Já existe Safra para esse ano.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Registra uma Safra, compartilhada por todas as Propriedades.",
        "tags": [
          "Safras"
        ]
      },
      "get": {
        "operationId": "SafrasController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SafrasDto"
                }
              }
            }
          }
        },
        "summary": "Lista as Safras, da mais recente para a mais antiga.",
        "tags": [
          "Safras"
        ]
      }
    },
    "/culturas": {
      "post": {
        "operationId": "CulturasController_acrescentar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcrescentarCulturaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturaDto"
                }
              }
            }
          },
          "400": {
            "description": "O nome informado não serve.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "O catálogo já tem essa espécie, ainda que escrita de outro jeito.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Acrescenta uma espécie ao catálogo.",
        "tags": [
          "Culturas"
        ]
      },
      "get": {
        "operationId": "CulturasController_listar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CulturasDto"
                }
              }
            }
          }
        },
        "summary": "Lista o catálogo em ordem alfabética.",
        "tags": [
          "Culturas"
        ]
      }
    }
  },
  "info": {
    "title": "Cadastro Rural",
    "description": "Cadastro de Produtores, Propriedades, Safras, Culturas e Plantios. Toda falha sai no formato Problem Details da RFC 9457.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CriarProdutorDto": {
        "type": "object",
        "properties": {
          "documento": {
            "type": "string",
            "minLength": 11,
            "maxLength": 18,
            "description": "CPF ou CNPJ, com ou sem máscara. Letra minúscula é recusada."
          },
          "nome": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Nome do Produtor."
          }
        },
        "required": [
          "documento",
          "nome"
        ]
      },
      "ProdutorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "nome": {
            "type": "string"
          },
          "documento": {
            "type": "string",
            "description": "Mascarado. Os dois últimos grupos aparecem, o resto não."
          },
          "tipoDeDocumento": {
            "type": "string",
            "enum": [
              "CPF",
              "CNPJ"
            ]
          }
        },
        "required": [
          "id",
          "nome",
          "documento",
          "tipoDeDocumento"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI que identifica o tipo do problema."
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "description": "A URI que sofreu a falha.",
            "type": "string"
          },
          "correlationId": {
            "description": "Liga a resposta às linhas de log da requisição.",
            "type": "string"
          },
          "codigo": {
            "description": "Código do erro de regra de negócio, quando a falha veio do domínio.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ]
      },
      "ProdutorDetalhadoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "nome": {
            "type": "string"
          },
          "documento": {
            "type": "string",
            "description": "Mascarado. Os dois últimos grupos aparecem, o resto não."
          },
          "tipoDeDocumento": {
            "type": "string",
            "enum": [
              "CPF",
              "CNPJ"
            ]
          },
          "propriedades": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "cidade": {
                  "type": "string"
                },
                "estado": {
                  "type": "string",
                  "description": "A sigla da unidade federativa."
                },
                "areaTotal": {
                  "type": "number",
                  "description": "Em hectares."
                },
                "areaAgricultavel": {
                  "type": "number",
                  "description": "Em hectares."
                },
                "areaDeVegetacao": {
                  "type": "number",
                  "description": "Em hectares."
                }
              },
              "required": [
                "id",
                "cidade",
                "estado",
                "areaTotal",
                "areaAgricultavel",
                "areaDeVegetacao"
              ]
            }
          }
        },
        "required": [
          "id",
          "nome",
          "documento",
          "tipoDeDocumento",
          "propriedades"
        ]
      },
      "CriarPropriedadeDto": {
        "type": "object",
        "properties": {
          "produtorId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "description": "O Produtor em nome de quem a Propriedade é registrada."
          },
          "cidade": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "estado": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "description": "A sigla da unidade federativa."
          },
          "areaTotal": {
            "type": "number",
            "minimum": 0,
            "description": "Em hectares, com até duas casas decimais."
          },
          "areaAgricultavel": {
            "type": "number",
            "minimum": 0,
            "description": "Em hectares, com até duas casas decimais."
          },
          "areaDeVegetacao": {
            "type": "number",
            "minimum": 0,
            "description": "Em hectares, com até duas casas decimais."
          }
        },
        "required": [
          "produtorId",
          "cidade",
          "estado",
          "areaTotal",
          "areaAgricultavel",
          "areaDeVegetacao"
        ]
      },
      "PropriedadeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "produtorId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "cidade": {
            "type": "string"
          },
          "estado": {
            "type": "string",
            "description": "A sigla da unidade federativa."
          },
          "areaTotal": {
            "type": "number",
            "description": "Em hectares."
          },
          "areaAgricultavel": {
            "type": "number",
            "description": "Em hectares."
          },
          "areaDeVegetacao": {
            "type": "number",
            "description": "Em hectares."
          }
        },
        "required": [
          "id",
          "produtorId",
          "cidade",
          "estado",
          "areaTotal",
          "areaAgricultavel",
          "areaDeVegetacao"
        ]
      },
      "PropriedadesPaginaDto": {
        "type": "object",
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "produtorId": {
                  "type": "string",
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                },
                "cidade": {
                  "type": "string"
                },
                "estado": {
                  "type": "string",
                  "description": "A sigla da unidade federativa."
                },
                "areaTotal": {
                  "type": "number",
                  "description": "Em hectares."
                },
                "areaAgricultavel": {
                  "type": "number",
                  "description": "Em hectares."
                },
                "areaDeVegetacao": {
                  "type": "number",
                  "description": "Em hectares."
                }
              },
              "required": [
                "id",
                "produtorId",
                "cidade",
                "estado",
                "areaTotal",
                "areaAgricultavel",
                "areaDeVegetacao"
              ]
            }
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Quantos registros existem ao todo."
          },
          "pagina": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "tamanho": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "itens",
          "total",
          "pagina",
          "tamanho"
        ]
      },
      "EditarPropriedadeDto": {
        "type": "object",
        "properties": {
          "cidade": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "estado": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "description": "A sigla da unidade federativa."
          },
          "areaTotal": {
            "type": "number",
            "minimum": 0,
            "description": "Em hectares, com até duas casas decimais."
          },
          "areaAgricultavel": {
            "type": "number",
            "minimum": 0,
            "description": "Em hectares, com até duas casas decimais."
          },
          "areaDeVegetacao": {
            "type": "number",
            "minimum": 0,
            "description": "Em hectares, com até duas casas decimais."
          }
        },
        "required": [
          "cidade",
          "estado",
          "areaTotal",
          "areaAgricultavel",
          "areaDeVegetacao"
        ]
      },
      "CriarSafraDto": {
        "type": "object",
        "properties": {
          "ano": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "Ano do ciclo agrícola, entre 1900 e 2100."
          }
        },
        "required": [
          "ano"
        ]
      },
      "SafraDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "ano": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "ano"
        ]
      },
      "SafrasDto": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "ano": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          },
          "required": [
            "id",
            "ano"
          ]
        }
      },
      "AcrescentarCulturaDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "description": "Nome da espécie, como Soja, Milho ou Café."
          }
        },
        "required": [
          "nome"
        ]
      },
      "CulturaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "nome": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome"
        ]
      },
      "CulturasDto": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "nome": {
              "type": "string"
            }
          },
          "required": [
            "id",
            "nome"
          ]
        }
      }
    }
  }
}